lab6: finished

+++ lab6/include/mini_uart.h
@@ -3,13 +3,15 @@
 
 #define BUFFER_SIZE 1024
 
+#include <stdint.h>
+
 void uart_init ( void );
 void uart_enable_interrupt( void );
 void uart_disable_interrupt( void );
 char uart_recv ( void );
 void uart_send ( char c );
 void uart_send_string(const char* str);
-void uart_hex(unsigned int d);
+void uart_hex(uint64_t d);
 void uart_irq_handler(void);
 char uart_async_recv( void );
 void uart_async_send_string(const char* buffer);
